Need to identify Questions & Options by an id; hit the friction point where content doesn't work
Just noticed that when you an edit a question or option, if there is
another question/option with the same prompt/content then they are
both edited. This is because we're identifying them by their
prompt/content :-P

It was incorrect to treat Question.prompt and Option as natural keys,
so let's start using ids, eh?

Also, bring an expression back onto a single line.